Make numba an optional dependency
numba pulls in llvmlite, together about 130 MB, for what is a single
module in this library. Add numpy implementations of the median and
mode filters and import the numba module only when it is available, so
the base installation no longer needs it.

The numpy implementations produce identical results. Both work on
strided views of the source buffer in slabs, which bounds the size of
the temporaries: the median uses the same sorting network as the numba
kernel, the mode counts matches pairwise. The median is about as fast
either way, the mode is roughly 5x slower than the numba kernel but
matches what the reshape-based implementation used to cost.

Tests run the parity checks over both paths against a reference that
does not need numba.
